Move get_auto_lims function to utils
